Remove wcc pre-attrs form Write and Commit Replies
Given that Ganesha need to make multiple calls to underlying filesystem
to perform write and then to get attributes, there is no way we can provide
accurate/atomic wcc pre-attrs. Because of this, the attr times on the client
side are not matching, resulting in invalidating client's cache. Removing
wcc pre-attrs in the Write and Commit replies, allows client to accurately
follow close-to-open semantics and keep the client cache sane.

Signed-off-by: Venkateswararao Jujjuri "<jvrao@linux.vnet.ibm.com>